pyproject.toml
[project]
name = "smol-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fsspec>=2024.12.0",
    "jupyterlab>=4.4.10",
    "langchain-text-splitters>=1.0.0",
    "neo4j-graphrag>=1.10.0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "redisvl>=0.10.0",
    "sentence-transformers>=5.1.2",
    "smolagents[toolkit]>=1.22.0",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
]
